pyproject.toml
[project]
name = "compiler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=14.0.0",
    "python-bebop>=1.3.2",
    "pyyaml>=6.0.0",
    "behave>=1.2.6",
]
